requirements.txt
tqdm
gym-retro
gym
scipy
numpy
gym
opencv-python
anyrl
tensorflow
scikit-learn